Did some refactoring. Just a little a tiny bit. Almost nothing changed. Also added VH_MAP to DataClsAdapter.
