+++ Android/app/src/main/java/com/example/boardgamesocial/DataViews/Adapters/DataClsAdapter.java
@@ -1,5 +1,6 @@
 package com.example.boardgamesocial.DataViews.Adapters;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -8,25 +9,36 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.boardgamesocial.DataClasses.DataClass;
+import com.example.boardgamesocial.DataClasses.Game;
+import com.example.boardgamesocial.DataClasses.Post;
 import com.example.boardgamesocial.DataViews.Adapters.ViewHolders.DataClsVH;
+import com.example.boardgamesocial.DataViews.Adapters.ViewHolders.PostVH;
+import com.example.boardgamesocial.DataViews.Adapters.ViewHolders.VHConstructor;
+import com.example.boardgamesocial.DataViews.DataClsVM;
 import com.example.boardgamesocial.R;
 
 import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
 import java.util.List;
-import java.util.function.Supplier;
+import java.util.Map;
+import java.util.Objects;
 
 public class DataClsAdapter <DC extends DataClass, VH extends DataClsVH<DC>> extends RecyclerView.Adapter<VH> {
     private List<DC> objectList;
-    private Class<VH> vhConstructor;
+//    private Constructor<VH> vhConstructor;
+    private Class<DC> cls;
 
-    public DataClsAdapter(List<DC> objectList, Class<VH> cls) {
+    public static final Map<Class<?>, VHConstructor> VH_MAP = new HashMap<Class<?>, VHConstructor>(){{
+        put(Post.class, PostVH::new);
+    }};
+
+    public DataClsAdapter(List<DC> objectList, Class<DC> cls) {
         this.objectList = objectList;
-        this.vhConstructor = cls;
+        this.cls = cls;
     }
 
-    public void setVhConstructor(Class<VH> cls) {
-        this.vhConstructor = cls;
+    public void setCls(Class<DC> cls) {
+        this.cls = cls;
     }
 
     public void setPosts(List<DC> objectList) {
@@ -36,13 +48,9 @@ public class DataClsAdapter <DC extends DataClass, VH extends DataClsVH<DC>> ext
     @NonNull
     @Override
     public VH onCreateViewHolder(ViewGroup parent, int viewType) {
-        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.item, parent,false);
-        try {
-            return vhConstructor.getConstructor().newInstance(itemView);
-        } catch (IllegalAccessException | InstantiationException | InvocationTargetException | NoSuchMethodException e) {
-            e.printStackTrace();
-            throw new IllegalArgumentException("Could not create instance of ViewHolder in DataClsAdapter");
-        }
+        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.post_item, parent,false);
+        Log.w("View", String.valueOf(itemView));
+        return (VH) Objects.requireNonNull(VH_MAP.get(cls)).create(itemView);
     }
 
     @Override
