+++ Android/app/src/main/java/com/example/boardgamesocial/DataViews/Adapters/ViewHolders/VHConstructor.java
@@ -1,0 +1,12 @@
+package com.example.boardgamesocial.DataViews.Adapters.ViewHolders;
+
+import android.view.View;
+
+import com.example.boardgamesocial.DataClasses.DataClass;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public interface VHConstructor {
+    DataClsVH<?> create(View view);
+}
