+++ Android/app/src/main/java/com/example/boardgamesocial/DataViews/DataClsVM.java
@@ -1,7 +1,9 @@
 package com.example.boardgamesocial.DataViews;
 
+import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModel;
 
 import com.example.boardgamesocial.DataClasses.DataClass;
