+++ Android/app/src/main/java/com/example/boardgamesocial/LoginAndSignUp/LoginAndSignUpActivity.java
@@ -1,12 +1,8 @@
-package com.example.boardgamesocial;
+package com.example.boardgamesocial.LoginAndSignUp;
 
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.navigation.NavController;
@@ -14,16 +10,8 @@ import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
-import com.example.boardgamesocial.API.RetrofitClient;
-import com.example.boardgamesocial.DataClasses.Token;
-import com.example.boardgamesocial.DataClasses.User;
+import com.example.boardgamesocial.R;
 import com.example.boardgamesocial.databinding.ActivityLoginAndSignUpBinding;
-
-import java.util.Objects;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 public class LoginAndSignUpActivity extends AppCompatActivity {
