+++ Android/app/src/main/java/com/example/boardgamesocial/MainApp/Fragments/AddPostFragment.java
@@ -1,4 +1,4 @@
-package com.example.boardgamesocial;
+package com.example.boardgamesocial.MainApp.Fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.boardgamesocial.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link SingleHostedGameFragment#newInstance} factory method to
+ * Use the {@link AddPostFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class SingleHostedGameFragment extends Fragment {
+public class AddPostFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class SingleHostedGameFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public SingleHostedGameFragment() {
+    public AddPostFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class SingleHostedGameFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment SingleHostedGameFragment.
+     * @return A new instance of fragment AddPostFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static SingleHostedGameFragment newInstance(String param1, String param2) {
-        SingleHostedGameFragment fragment = new SingleHostedGameFragment();
+    public static AddPostFragment newInstance(String param1, String param2) {
+        AddPostFragment fragment = new AddPostFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class SingleHostedGameFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_single_hosted_game, container, false);
+        return inflater.inflate(R.layout.fragment_add_post, container, false);
     }
 }
