+++ Android/app/src/main/java/com/example/boardgamesocial/MainApp/Fragments/EventAttendeesFragment.java
@@ -1,4 +1,4 @@
-package com.example.boardgamesocial;
+package com.example.boardgamesocial.MainApp.Fragments;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.boardgamesocial.R;
 
 /**
  * A simple {@link Fragment} subclass.
