+++ Android/app/src/main/java/com/example/boardgamesocial/MainApp/Fragments/ProfileFragment.java
@@ -1,30 +1,37 @@
-package com.example.boardgamesocial;
+package com.example.boardgamesocial.MainApp.Fragments;
 
 import android.os.Bundle;
 
+import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+
+import androidx.annotation.NonNull;
+import androidx.navigation.fragment.NavHostFragment;
+
+import com.example.boardgamesocial.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link SinglePostFragment#newInstance} factory method to
+ * Use the {@link ProfileFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class SinglePostFragment extends Fragment {
+public class ProfileFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
 
-    // TODO: Rename and change types of parameters
+//    // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
 
-    public SinglePostFragment() {
+    public ProfileFragment() {
         // Required empty public constructor
     }
 
@@ -32,13 +39,11 @@ public class SinglePostFragment extends Fragment {
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment SinglePostFragment.
+     * returns a new instance of fragment ProfileFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static SinglePostFragment newInstance(String param1, String param2) {
-        SinglePostFragment fragment = new SinglePostFragment();
+    public static ProfileFragment newInstance(String param1, String param2) {
+        ProfileFragment fragment = new ProfileFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +64,15 @@ public class SinglePostFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_single_post, container, false);
+        return inflater.inflate(R.layout.fragment_profile, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        Button buttonEditProfile = view.findViewById(R.id.profile_btn_edit);
+
+        buttonEditProfile.setOnClickListener(view1 -> NavHostFragment.findNavController(ProfileFragment.this)
+                .navigate(R.id.action_profileFragment_to_editProfileFragment));
     }
 }
