--- Android/app/src/main/java/com/example/boardgamesocial/SignUpActivity.java
@@ -1,33 +1,0 @@
-package com.example.boardgamesocial;
-
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.example.boardgamesocial.API.RetrofitClient;
-import com.example.boardgamesocial.DataClasses.User;
-
-import java.util.Objects;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class SignUpActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sign_up);
-    }
-
-    public static Intent getIntent(Context context) {
-        return new Intent(context, SignUpActivity.class);
-    }
-}
