--- Android/app/src/main/res/layout/activity_games_collection.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".GamesCollectionActivity">
-
-    <!--  https://youtu.be/IfgV9WzYOn8?t=443  -->
-
-    <include
-        android:id="@+id/include_main_app_fragment"
-        layout="@layout/game_items"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:padding="@dimen/main_app_activity_padding"
-        tools:layout_editor_absoluteX="0dp" />
-
-    <com.google.android.material.bottomappbar.BottomAppBar
-        android:id="@+id/bottom_app_bar_games_collection"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:backgroundTint="@color/very_light_tangelo"
-        app:fabAlignmentMode="end"
-        app:itemIconTint="@color/ice_blue"
-        app:itemTextColor="@color/blue_yonder"
-        app:menu="@menu/navbar_options" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/home_post_fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:backgroundTint="@color/coral_reef"
-        app:layout_anchor="@id/bottom_app_bar"
-        app:srcCompat="@drawable/ic_baseline_add_24" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
