+++ Android/app/src/main/res/layout/activity_login_and_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".LoginAndSignUpActivity">
+    tools:context=".LoginAndSignUp.LoginAndSignUpActivity">
 
 
     <include
