+++ Android/app/src/main/res/layout/fragment_add_post.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddPostFragment">
+    tools:context=".MainApp.Fragments.AddPostFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
