+++ Android/app/src/main/res/layout/fragment_edit_profile.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfileFragment">
+    tools:context=".MainApp.Fragments.EditProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
