+++ Android/app/src/main/res/layout/fragment_event_attendees.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventAttendeesFragment">
+    tools:context=".MainApp.Fragments.EventAttendeesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
