+++ Android/app/src/main/res/layout/fragment_events.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventsFragment">
+    tools:context=".MainApp.Fragments.EventsFragment">
 
     <TextView
         android:id="@+id/textview_second"
