+++ Android/app/src/main/res/layout/fragment_friend_chat.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FriendChatFragment">
+    tools:context=".MainApp.Fragments.FriendChatFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
