+++ Android/app/src/main/res/layout/fragment_game_collection.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameCollectionFragment">
+    tools:context=".MainApp.Fragments.GameCollectionFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
