+++ Android/app/src/main/res/layout/fragment_home_post.xml
@@ -11,13 +11,13 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" >
 
-        <LinearLayout
+        <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/postFeed_recyclerView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
-            tools:listitem="@layout/item"
+            tools:listitem="@layout/post_item"
             android:orientation="vertical">
-        </LinearLayout>
+        </androidx.recyclerview.widget.RecyclerView>
     </RelativeLayout>
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
