+++ Android/app/src/main/res/layout/fragment_hosted_games.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HostedGamesFragment">
+    tools:context=".MainApp.Fragments.HostedGamesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
