+++ Android/app/src/main/res/layout/fragment_sign_up.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignUpFragment">
+    tools:context=".LoginAndSignUp.Fragments.SignUpFragment">
 
     <EditText
         android:id="@+id/et_signup_firstname"
