+++ Android/app/src/main/res/layout/fragment_single_event.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SingleEventFragment">
+    tools:context=".MainApp.Fragments.SingleEventFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
