+++ Android/app/src/main/res/layout/fragment_single_post.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SinglePostFragment">
+    tools:context=".MainApp.Fragments.SinglePostFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
