+++ Android/app/src/main/res/layout/fragment_user_friends.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserFriendsFragment">
+    tools:context=".MainApp.Fragments.UserFriendsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
