+++ Android/app/src/main/res/layout/game_item.xml
