+++ Android/app/src/main/res/layout/post_item.xml
