+++ Android/app/src/main/res/navigation/nav_login_and_sign_up.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/LoginFragment"
-        android:name="com.example.boardgamesocial.LoginFragment"
+        android:name="com.example.boardgamesocial.LoginAndSignUp.Fragments.LoginFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_login">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SignUpFragment"
-        android:name="com.example.boardgamesocial.SignUpFragment"
+        android:name="com.example.boardgamesocial.LoginAndSignUp.Fragments.SignUpFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_sign_up">
